refactor: deduplicate Reval Person code
Reducing LoC and redundancy

TIS21-5812: Add nightly job for Funding Status

TIS21-6044: Address code duplication smells
