+++ src/main/java/uk/nhs/tis/sync/job/reval/RevalCurrentPlacementSyncJob.java
@@ -2,38 +2,31 @@ package uk.nhs.tis.sync.job.reval;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
-import javax.persistence.EntityManager;
 import net.javacrumbs.shedlock.core.SchedulerLock;
-import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.context.annotation.Profile;
 import org.springframework.jmx.export.annotation.ManagedOperation;
 import org.springframework.jmx.export.annotation.ManagedResource;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
-import uk.nhs.tis.sync.job.PersonDateChangeCaptureSyncJobTemplate;
 import uk.nhs.tis.sync.message.publisher.RabbitMqTcsRevalTraineeUpdatePublisher;
-import uk.nhs.tis.sync.model.EntityData;
 
 /**
- * Get personIds whose current placement changes nightly. And sends messages to rabbitMq
- * for tcs to fetch
+ * Get personIds whose current placement changed overnight, i.e. ended yesterday or started today.
+ * This sends messages for further processing, currently tcs to fetch and forward relevant data.
  */
 @Profile("!nimdta")
 @Component
 @ManagedResource(objectName = "sync.mbean:name=RevalCurrentPlacementSyncJob",
     description = "Job message personIds if their placement(s) started/ended")
-public class RevalCurrentPlacementSyncJob extends PersonDateChangeCaptureSyncJobTemplate<Long> {
+public class RevalCurrentPlacementSyncJob extends RevalPersonChangedJobTemplate {
+
   private static final String BASE_QUERY =
       "SELECT DISTINCT traineeId FROM Placement" + " WHERE traineeId > :lastPersonId"
           + " AND (dateFrom = ':today' OR dateTo = ':yesterday')"
           + " ORDER BY traineeId LIMIT :pageSize";
-  private final RabbitMqTcsRevalTraineeUpdatePublisher rabbitMqPublisher;
 
   public RevalCurrentPlacementSyncJob(RabbitMqTcsRevalTraineeUpdatePublisher rabbitMqPublisher) {
-    this.rabbitMqPublisher = rabbitMqPublisher;
+    super(rabbitMqPublisher);
   }
 
   @Override
@@ -49,22 +42,6 @@ public class RevalCurrentPlacementSyncJob extends PersonDateChangeCaptureSyncJob
       description = "send personIds to tcs for reval current placement sync")
   public void revalCurrentPlacementSyncJob() {
     runSyncJob(null);
-  }
-
-  @Override
-  protected int convertData(Set<Long> entitiesToSave, List<EntityData> entityData,
-                            EntityManager entityManager) {
-    entitiesToSave.addAll(
-        entityData.stream().map(EntityData::getEntityId).collect(Collectors.toList()));
-    return 0;
-  }
-
-  @Override
-  protected void handleData(Set<Long> dataToSave, EntityManager entityManager) {
-    if (CollectionUtils.isNotEmpty(dataToSave)) {
-      rabbitMqPublisher.publishToBroker(
-          dataToSave.stream().map(String::valueOf).collect(Collectors.toSet()));
-    }
   }
 
   @Override
