+++ src/main/java/uk/nhs/tis/sync/job/reval/RevalCurrentPmSyncJob.java
@@ -2,20 +2,13 @@ package uk.nhs.tis.sync.job.reval;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
-import javax.persistence.EntityManager;
 import net.javacrumbs.shedlock.core.SchedulerLock;
-import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.context.annotation.Profile;
 import org.springframework.jmx.export.annotation.ManagedOperation;
 import org.springframework.jmx.export.annotation.ManagedResource;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
-import uk.nhs.tis.sync.job.PersonDateChangeCaptureSyncJobTemplate;
 import uk.nhs.tis.sync.message.publisher.RabbitMqTcsRevalTraineeUpdatePublisher;
-import uk.nhs.tis.sync.model.EntityData;
 
 /**
  * Get personIds whose current programmeMembership changes nightly. And sends messages to rabbitMq
@@ -25,16 +18,15 @@ import uk.nhs.tis.sync.model.EntityData;
 @Component
 @ManagedResource(objectName = "sync.mbean:name=RevalCurrentPmSyncJob",
     description = "Job message personIds if their programme membership(s) started/ended")
-public class RevalCurrentPmSyncJob extends PersonDateChangeCaptureSyncJobTemplate<Long> {
+public class RevalCurrentPmSyncJob extends RevalPersonChangedJobTemplate {
 
   private static final String BASE_QUERY =
       "SELECT DISTINCT personId FROM ProgrammeMembership" + " WHERE personId > :lastPersonId"
           + " AND (programmeEndDate = ':endDate' OR programmeStartDate = ':startDate')"
           + " ORDER BY personId LIMIT :pageSize";
-  private final RabbitMqTcsRevalTraineeUpdatePublisher rabbitMqPublisher;
 
   public RevalCurrentPmSyncJob(RabbitMqTcsRevalTraineeUpdatePublisher rabbitMqPublisher) {
-    this.rabbitMqPublisher = rabbitMqPublisher;
+    super(rabbitMqPublisher);
   }
 
   @Override
@@ -50,22 +42,6 @@ public class RevalCurrentPmSyncJob extends PersonDateChangeCaptureSyncJobTemplat
       description = "send personIds to tcs for reval current programmeMembership sync")
   public void revalCurrentPmSyncJob() {
     runSyncJob(null);
-  }
-
-  @Override
-  protected int convertData(Set<Long> entitiesToSave, List<EntityData> entityData,
-                            EntityManager entityManager) {
-    entitiesToSave.addAll(
-        entityData.stream().map(EntityData::getEntityId).collect(Collectors.toList()));
-    return 0;
-  }
-
-  @Override
-  protected void handleData(Set<Long> dataToSave, EntityManager entityManager) {
-    if (CollectionUtils.isNotEmpty(dataToSave)) {
-      rabbitMqPublisher.publishToBroker(
-          dataToSave.stream().map(String::valueOf).collect(Collectors.toSet()));
-    }
   }
 
   @Override
